+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Redirect automatically to project#index after the user authentication, when there is only one team on enrollments
 - **Increase decimal precision from Stories position**
 - Change story controls to react component
+- Links to stories within the same project can be added to a story description.
 
 ## [1.1.3] - 2017-03-30
 ### Changed